Редактирование 7 модуль
