+++ js/main.js
@@ -1,3 +1,4 @@
-import './templates.js';
+import { renderPictures } from './photoCreationTemplate.js';
+import { pictures } from './data.js';
 
-
+renderPictures(pictures);
